requirements-base.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-multipart==0.0.6
httpx==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
pdf2image==1.16.3
python-dateutil==2.8.2
pytz==2023.3
babel==2.12.1
pytest==8.3.5
pytest-asyncio==0.23.5